auth-server/requirements.txt
flask==3.0.0
stripe==7.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
